AMB-813 delete old flask app
