--- docker/receiver/src_old/routes/examples/appointment/GET-success.json
@@ -1,9 +1,0 @@
-{
-  "resourceType": "Bundle",
-  "meta": {
-    "profile": [
-      "https://simplifier.net/guide/UKCoreDevelopment2/ProfileUKCore-Bundle"
-    ],
-    "type": "searchset"
-  }
-}
