--- docker/receiver/src_old/routes/examples/appointment/POST-success.txt
@@ -1,1 +1,0 @@
-c3f6145e-1a26-4345-b3f2-dccbcba62049
