--- docker/receiver/src_old/routes/examples/appointment/id/GET-success.json
@@ -1,8 +1,0 @@
-{
-  "resourceType": "Appointment",
-  "meta": {
-    "profile": [
-      "https://www.hl7.org/fhir/appointment"
-    ]
-  }
-}
