--- docker/receiver/src_old/routes/examples/service_request/id/PUT-success.json
@@ -1,8 +1,0 @@
-{
-  "resourceType": "ServiceRequest",
-  "meta": {
-    "profile": [
-      "https://simplifier.net/guide/UKCoreDevelopment2/ProfileUKCore-ServiceRequest"
-    ]
-  }
-}
